// src/state/board/types.ts
export enum types {
  loadData = "LOAD_DATA",
  columnNameChange = "COLUMN_NAME_CHANGE",
  toggleAddCardField = "TOGGLE_ADD_CARD_FIELD",
  pushNewCard = "PUSH_NEW_CARD",
  cardDescriptionChange = "CARD_DESCRIPTION_CHANGE",
  cardNameChange = "CARD_NAME_CHANGE",
  addNewComment = "ADD_NEW_COMENT",
  commentEditSave = "COMMENT_EDIT_SAVE",
  commentDelite = "COMMENT_DELITE",
  deleteCard = "DELETE_CARD",
}

export default types;